Imgui was removed. First tool - Box - was added to editor

+++ core/Entity.cpp
@@ -1,3 +1,3 @@
 #include "Entity.h"
 
-int Entity::nextId = 0;
+size_t Entity::nextId = 0;
